rename module_1a -> time_map, remove module_1
