--- grass/variables/defaults/module_1.ps_param_1
@@ -1,45 +1,0 @@
-maploc 0 2
-
-border
-  color grey
-  width 1
-  end
-
-mapinfo
-  where 1.5 0
-  font Courier
-  fontsize 12
-  color brown
-  end
-
-vlines lines
-  where waterway = 'stream' OR waterway = 'drain' OR waterway = 'canal'
-  color blue
-  width 0.2
-  masked y
-  end
-
-vlines lines
-  where waterway = 'river'
-  color #ccccff
-  width 3
-  masked y
-  end
-
-vlines m1_stricken_area
-  color #000000
-  width 0.2
-  masked n
-  end
-
-raster m1_time_map
-
-colortable y
-  where 0 0
-  cols 1
-  width 3
-  font Helvetica
-  end
-
-paper a3
-  end
