--- grass/variables/defaults/module_1.ps_param_2
@@ -1,77 +1,0 @@
-
-maploc 0 2
-
-border
-  color grey
-  width 1
-  end
-
-mapinfo
-  where 1.5 0
-  font Courier
-  fontsize 12
-  color brown
-  end
-
-vlines lines
-  where highway = 'trunk' OR highway = 'trunk link' OR 'primary' OR highway = 'primary link' OR highway = 'secondary' OR highway = 'secondary link' OR highway = 'tertiary' OR highway = 'tertiary link' OR highway = 'residential' OR highway = 'service' OR highway = 'truck'
-  color grey
-  width 0.2
-  masked y
-  end
-
-vlines lines
-  where waterway = 'stream' OR waterway = 'drain' OR waterway = 'canal'
-  color blue
-  width 0.2
-  masked y
-  end
-
-vlines lines
-  where waterway = 'river'
-  color #aaaaff
-  width 4
-  masked y
-  end
-
-raster m1_time_map_interpolated
-
-vlines m1_stricken_area
-  color #000000
-  width 0.7
-  masked n
-  end
-
-vpoints m1_from_points
-  symbol basic/cross3
-  size 6
-  width 0.5
-  color black
-  fcolor yellow
-  end
-
-vpoints m1_via_points
-  symbol basic/cross3
-  size 6
-  width 0.5
-  color black
-  fcolor #ff00ff
-  end
-
-vpoints m1_to_points
-  symbol basic/cross3
-  size 6
-  width 0.5
-  color black
-  fcolor #ff0000
-  end
-
-colortable y
-  where 0 0
-  cols 1
-  width 3
-  font Helvetica
-  end
-
-paper a3
-  end
