+++ webapp/views/launch/index.pug
@@ -27,7 +27,7 @@ html
             li(class="nav-item d-flex align-items-center")
               span(class="d-inline-block align-middle m-2") Modules:
               select(id='launch-module-menu' class='custom-select custom-select-lg mr-2')
-                option(selected value='module_1a') Calculate time map
+                option(selected value='time_map') Calculate time map
                 option(value='module_2') Query area
               button(class='btn btn-success btn-lg mr-2' onclick='launchModule()') ▷&nbsp;Run
               button(class='btn btn-light btn-lg' id="result-btn" onclick='showResults()') Results
@@ -82,7 +82,7 @@ html
           placement: 'bottom',
           title: function () {
             switch ($(this).val()) {
-              case 'module_1a': return 'The ‘Calculate time map’ module creates a map showing the time it takes to reach any part of the selected area from a given start point. Optionally, via points and/or affected areas with reduced speed can be defined.';
+              case 'time_map': return 'The ‘Calculate time map’ module creates a map showing the time it takes to reach any part of the selected area from a given start point. Optionally, via points and/or affected areas with reduced speed can be defined.';
               case 'module_2': return 'The ‘Query area’ module lets you filter map features based on a given set of constraints. For example, if a housing map layer is queried based on the household population and monthly income, only the houses within the given range will be returned.';
             }
           },
